With this cset, I only get one warning while compiling bk:
dataheap.c: In function ‘datamap’:
dataheap.c:365: warning: label ‘nopage’ defined but not used

I assume that one is temporary?

bk: 50060919XUpLV-EX9Zvg39ax3iAi0w

+++ src/libc/utils/fopen_vzip.c
@@ -348,8 +348,8 @@ err:		errno = EINVAL;
 		if (offset == 0) {
 			break;
 		} else if (offset < 0) {
-			fprintf(stderr, "fgzip: fseek to invalid offset %ld\n",
-				fz->offset);
+			fprintf(stderr, "fgzip: fseek to invalid offset %llu\n",
+			  (u64)fz->offset);
 			goto err;
 		}
 		offset -= sz->usz;
